M2 2592

+++ scripts/clean.sh
@@ -14,28 +14,10 @@ rm -rf $OS_BUILD/src/build
 echo "Cleaning src dist dir"
 rm -rf $OS_BUILD/src/dist
 
-echo "Cleaning bower components"
-rm -rf $OS_BUILD/src/bower_components
-
-echo "Cleaning node modules"
-rm -rf $OS_BUILD/src/node_modules
+if [ "$OS_DO_BUILD" = true ] ; then
+    echo "Cleaning node modules"
+    rm -rf $OS_BUILD/src/node_modules
+fi
 
 echo "Cleaning package lock "
 rm -rf $OS_BUILD/src/package-lock.json
-
-echo "Restore original package.json "
-if [ -f $OS_BUILD/src/package.json.backup ]; then
-  rm -rf $OS_BUILD/src/package.json
-  mv $OS_BUILD/src/package.json.backup $OS_BUILD/src/package.json
-fi
-
-echo "Restoring original gulp gulpfile"
-# check if gulp old gulpfile exist
-if [ -f $OS_BUILD/src/gulpfile.js.old ]; then
-  # restore original project config file
-  rm -rf $OS_BUILD/src/gulpfile.js
-  mv $OS_BUILD/src/gulpfile.js.old $OS_BUILD/src/gulpfile.js
-fi
-
-echo "Cleaning gulp project files"
-rm -rf $OS_BUILD/src/gulp/debug.js
